+++ tools/marching_cubes_reconstruction.cpp
@@ -149,11 +149,11 @@ main (int argc, char** argv)
 
 
   // Command line parsing
-  double leaf_size = default_leaf_size;
+  float leaf_size = default_leaf_size;
   parse_argument (argc, argv, "-leaf", leaf_size);
   print_info ("Using a leaf size of: "); print_value ("%f\n", leaf_size);
 
-  double iso_level = default_iso_level;
+  float iso_level = default_iso_level;
   parse_argument (argc, argv, "-iso", iso_level);
   print_info ("Setting an iso level of: "); print_value ("%f\n", iso_level);
